requirements.txt
pyrofork
tgcrypto
python-dotenv